Changed Talon's and Erik's Holonomic Field Centric to be potentially easier to read, and deleted the comments as we should rewrite them anyways.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOpTemplates/HolonomicOpModes/Holonomic_Hardware.java
@@ -23,7 +23,7 @@ public class Holonomic_Hardware {
     public DcMotor fleft, fright, bleft, bright;
     public BNO055IMU gyro;
     public double heading;
-    public float dp = .2f; //Drive Power (range = 0-1)
+    public double dp = .2f; //Drive Power (range = 0-1)
     private HardwareMap hwMap;
     private Telemetry telemetry;
